requirements.txt
python-telegram-bot
python-dotenv
aiosqlite
requests
aiohttp